# public/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FileFlow Project API NEW",
    "version": "1.0.0",
    "description": "API documentation for the FileFlow Project. All endpoints are synchronized with backend routes as of September 27, 2025."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/v1/analytics/trends/users": {
      "get": {
        "summary": "User Trends Analytics",
        "description": "Returns user trends analytics data.",
        "responses": {
          "200": {
            "description": "User trends analytics returned successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
...
    "/v1/admin/users": {
      "get": {
        "summary": "List All Users (Admin)",
        "description": "Returns a list of all users. Admin access required.",
        "responses": {
          "200": {
            "description": "List of users returned successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      },  "post": {
      "summary": "Bulk User Action (Admin)",
      "description": "Performs a bulk action on users. Admin access required.",
      "requestBody": {
        "content": {
          "application/json": {
            "schema": {
              "type": "object"
            }
          }
        }
      },
      "responses": {
        "200": {
          "description": "Bulk action completed successfully."
        },
        "400": {
          "description": "Validation error."
        },
        "401": {
          "description": "Unauthorized."
        },
        "403": {
          "description": "Forbidden."
        },
        "500": {
          "description": "Internal server error."
        }
      }
    }
    },
    "/v1/admin/users/{userId}": {
      "get": {
        "summary": "Get User by ID (Admin)",
        "description": "Returns details for a specific user by userId. Admin access required.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User details returned successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "User not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }, "patch": {
      "summary": "Update User (Admin)",
      "description": "Updates a user's details. Admin access required.",
      "parameters": [
        {
          "name": "userId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "requestBody": {
        "content": {
          "application/json": {
            "schema": {
              "type": "object"
            }
          }
        }
      },
      "responses": {
        "200": {
          "description": "User updated successfully."
        },
        "400": {
          "description": "Validation error."
        },
        "401": {
          "description": "Unauthorized."
        },
        "403": {
          "description": "Forbidden."
        },
        "404": {
          "description": "User not found."
        },
        "500": {
          "description": "Internal server error."
        }
      }
    },
    "delete": {
      "summary": "Delete User (Admin)",
      "description": "Deletes a user. Admin access required.",
      "parameters": [
        {
          "name": "userId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "responses": {
        "200": {
          "description": "User deleted successfully."
        },
        "401": {
          "description": "Unauthorized."
        },
        "403": {
          "description": "Forbidden."
        },
        "404": {
          "description": "User not found."
        },
        "500": {
          "description": "Internal server error."
        }
      }
    }
    },
...
    "/v1/admin/files": {
      "get": {
        "summary": "List All Files (Admin)",
        "description": "Returns a list of all files. Admin access required.",
        "responses": {
          "200": {
            "description": "List of files returned successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/v1/admin/files/{fileId}": {
      "get": {
        "summary": "Get File by ID (Admin)",
        "description": "Returns details for a specific file. Admin access required.",
        "parameters": [
          {
            "name": "fileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File details returned successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "File not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      },
      "patch": {
        "summary": "Update File (Admin)",
        "description": "Updates a file's metadata. Admin access required.",
        "parameters": [
          {
            "name": "fileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File updated successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "File not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      },
      "delete": {
        "summary": "Delete File (Admin)",
        "description": "Deletes a file. Admin access required.",
        "parameters": [
          {
            "name": "fileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File deleted successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "File not found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
...
    "/v1/admin/security-dashboard/users/{userId}": {
      "get": {
        "summary": "Get Security Metrics for User (Admin)",
        "description": "Returns security metrics for a specific user. Admin access required.",
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Security metrics returned successfully." },
          "401": { "description": "Unauthorized." },
          "403": { "description": "Forbidden." },
          "404": { "description": "User not found." },
          "500": { "description": "Internal server error." }
        }
      }
    },
...
  "/v1/profile/storage/stats": {
    "get": {
      "summary": "Get User Storage Stats",
      "description": "Returns storage usage statistics for the authenticated user.",
      "responses": {
        "200": {
          "description": "Storage stats returned successfully."
        },
        "401": {
          "description": "Unauthorized."
        },
        "500": {
          "description": "Internal server error."
        }
      }
    }
  },
  "/v1/profile/ip-whitelist/enable": {
    "post": {
      "summary": "Enable IP Whitelist",
      "description": "Enables IP whitelisting for the authenticated user.",
      "responses": {
        "200": {
          "description": "IP whitelist enabled successfully."
        },
        "401": {
          "description": "Unauthorized."
        },
        "500": {
          "description": "Internal server error."
        }
      }
    }
  },
  "/v1/profile/ip-whitelist/disable": {
    "post": {
      "summary": "Disable IP Whitelist",
      "description": "Disables IP whitelisting for the authenticated user.",
      "responses": {
        "200": {
          "description": "IP whitelist disabled successfully."
        },
        "401": {
          "description": "Unauthorized."
        },
        "500": {
          "description": "Internal server error."
        }
      }
    }
  },
  "/v1/profile/ip-whitelist/add": {
    "post": {
      "summary": "Add IP to Whitelist",
      "description": "Adds an IP address to the user's whitelist.",
      "requestBody": {
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "ip": {
                  "type": "string"
                }
              },
              "required": [
                "ip"
              ]
            }
          }
        }
      },
      "responses": {
        "200": {
          "description": "IP added to whitelist successfully."
        },
        "401": {
          "description": "Unauthorized."
        },
        "500": {
          "description": "Internal server error."
        }
      }
    }
  },
  "/v1/profile/ip-whitelist/remove": {
    "post": {
      "summary": "Remove IP from Whitelist",
      "description": "Removes an IP address from the user's whitelist.",
      "requestBody": {
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "ip": {
                  "type": "string"
                }
              },
              "required": [
                "ip"
              ]
            }
          }
        }
      },
      "responses": {
        "200": {
          "description": "IP removed from whitelist successfully."
        },
        "401": {
          "description": "Unauthorized."
        },
        "500": {
          "description": "Internal server error."
        }
      }
    }
  },
  "/v1/profile/files/bulk-operations": {
    "post": {
      "summary": "Bulk File Operations",
      "description": "Performs bulk operations on user files (delete, move, etc.).",
      "requestBody": {
        "content": {
          "application/json": {
            "schema": {
              "type": "object"
            }
          }
        }
      },
      "responses": {
        "200": {
          "description": "Bulk operation completed successfully."
        },
        "400": {
          "description": "Validation error."
        },
        "401": {
          "description": "Unauthorized."
        },
        "500": {
          "description": "Internal server error."
        }
      }
    }
  },
  "/v1/profile/me": {
    "get": {
      "summary": "Get Current User Profile",
      "description": "Returns the profile of the currently authenticated user.",
      "responses": {
        "200": {
          "description": "User profile returned successfully."
        },
        "401": {
          "description": "Unauthorized."
        },
        "500": {
          "description": "Internal server error."
        }
      }
    }
  },
  "/v1/profile": {
    "get": {
      "summary": "Get User Profile",
      "description": "Returns the profile of the authenticated user.",
      "responses": {
        "200": {
          "description": "User profile returned successfully."
        },
        "401": {
          "description": "Unauthorized."
        },
        "500": {
          "description": "Internal server error."
        }
      }
    },
    "patch": {
      "summary": "Update User Profile",
      "description": "Updates the profile of the authenticated user.",
      "requestBody": {
        "content": {
          "application/json": {
            "schema": {
              "type": "object"
            }
          }
        }
      },
      "responses": {
        "200": {
          "description": "User profile updated successfully."
        },
        "400": {
          "description": "Validation error."
        },
        "401": {
          "description": "Unauthorized."
        },
        "500": {
          "description": "Internal server error."
        }
      }
    }
  },
...
}
}